+++ packages/webcert/src/store/certificate/certificateMiddleware.ts
@@ -2,6 +2,7 @@ import {
   Certificate,
   CertificateDataElement,
   CertificateDataValidationType,
+  CertificateSignStatus,
   CertificateStatus,
   getCertificateToSave,
   SigningMethod,
@@ -22,7 +23,6 @@ import {
   autoSaveCertificateStarted,
   autoSaveCertificateSuccess,
   certificateApiGenericError,
-  CertificateSignStatus,
   complementCertificate,
   complementCertificateStarted,
   complementCertificateSuccess,
